Move generateCSV function up, nit fixes
