+++ backend/typescript/services/interfaces/campService.ts
@@ -64,7 +64,7 @@ interface ICampService {
   ): Promise<Array<CampSessionDTO>>;
 
   /**
-   * Get all campers associated with camp session of id campSessionId
+   * Get all campers associated with camp session of id campSessionId for CSV file generation
    * @param campSessionId camp's session id
    * @returns array of CamperCSVInfoDTO object containing campers information
    * @throws Error if camper retrieval fails
