+++ frontend/src/components/pages/CampOverview/CampersTable/CampersTable.tsx
@@ -34,43 +34,26 @@ import ViewCamperModal from "../ViewCamperModal/index";
 import { FormQuestion } from "../../../../types/CampsTypes";
 import RemoveCamperModal from "../RemoveCamperModal/index";
 
-import CampsAPIClient from "../../../../APIClients/CampsAPIClient";
-
-import { downloadCSV } from "../../../../utils/CSVUtils";
-
-const ExportButton = ({campSessionId, campSessionCSVFilename}: {campSessionId: string, campSessionCSVFilename: string}): JSX.Element => {
-  const toast = useToast();
-  
-  const generateCsv = async () => {
-    const csvResponse = await CampsAPIClient.getCampSessionCsv(campSessionId);
-    if (csvResponse !== "ERROR") {
-      downloadCSV(csvResponse, campSessionCSVFilename);
-    } else {
-      toast({
-        description: `An error occurred while exporting the CSV. Please try again.`,
-        status: "error",
-        variant: "subtle",
-        duration: 3000,
-      });
-    }
-  }
-  
+const ExportButton = ({generateCsv}: {generateCsv: () => void}): JSX.Element => {  
   return (
-    <Button
-      leftIcon={<DownloadIcon />}
-      aria-label="Export SVG"
-      type="submit"
-      background="background.grey.200"
-      color="primary.green.100"
-      border="2px"
-      borderColor="primary.green.100"
-      borderRadius="5px"
-      minWidth="-webkit-fit-content"
-      fontSize={textStyles.bodyRegular.fontSize}
-      onClick={() => generateCsv()}
-    >
-      Export as .csv
-    </Button>
+    <Box>
+      <Button
+        ml="30px"
+        leftIcon={<DownloadIcon />}
+        aria-label="Export SVG"
+        type="submit"
+        background="background.grey.200"
+        color="primary.green.100"
+        border="2px"
+        borderColor="primary.green.100"
+        borderRadius="5px"
+        minWidth="-webkit-fit-content"
+        fontSize={textStyles.bodyRegular.fontSize}
+        onClick={() => generateCsv()}
+      >
+        Export as .csv
+      </Button>
+    </Box>
   );
 };
 
@@ -79,15 +62,13 @@ const CampersTable = ({
   campSessionCapacity,
   formQuestions,
   handleRefetch,
-  campSessionId,
-  campSessionCSVFilename,
+  generateCsv,
 }: {
   campers: Camper[];
   campSessionCapacity: number;
   formQuestions: FormQuestion[];
   handleRefetch: () => void;
-  campSessionId: string;
-  campSessionCSVFilename: string;
+  generateCsv: () => void;
 }): JSX.Element => {
   const [search, setSearch] = React.useState("");
   const [selectedFilter, setSelectedFilter] = React.useState<Filter>(
@@ -204,7 +185,7 @@ const CampersTable = ({
                 </Tag>
               );
             })}
-            <ExportButton campSessionId={campSessionId} campSessionCSVFilename={campSessionCSVFilename}/>
+            <ExportButton generateCsv={generateCsv}/>
           </HStack>
 
           <Table
